requirements.txt
langchain>=0.3.23,<0.4.0
langgraph>=0.3.31,<0.4.0
langchain-openai>=0.3.14
langchain-anthropic>=0.3.12
langchain-community>=0.3.21,<0.4.0
python-dotenv>=1.1.0
openai>=1.75.0
anthropic>=0.49.0
pytest>=8.3.0
pytest-asyncio>=0.26.0
pytest-cov>=4.1.0
mkdocstrings-python>=1.16.10
python-chess>=1.999

# Celery and related dependencies
celery==5.5.1
redis==5.0.1
flower==2.0.1
msgpack==1.0.7